Refactor route handler into a controller (will make things cleaner later)
