+++ src/routes/index.ts
@@ -1,15 +1,7 @@
 import express from 'express';
 const router: express.Router = express.Router();
+import { indexController } from '../controllers/indexController';
 
-router.get('/', (req, res) => {
-  console.log('Hello world received a request.');
-
-  const name: string = req.ip || 'World';
-  res.send(typeAnnotationTest(name));
-});
-
-function typeAnnotationTest(name: string): string {
-  return `Hello ${name}!`;
-}
+router.get('/', indexController.index);
 
 export const index = router;
